requirements.txt
numpy==2.2.4
tqdm==4.67.1
PyYAML==6.0.2
scipy==1.15.2